refactor: remove intermediate context to prevent flaky errors
