+++ packages/app-builder/src/services/editor/index.ts
@@ -1,3 +1,2 @@
 export * from './coerceToConstantEditableAstNode';
-export * from './current-scenario-iteration';
 export * from './editor-mode';
