+++ packages/app-builder/src/services/validation/index.ts
@@ -1,3 +1,2 @@
-export * from './current-scenario-validation';
 export * from './scenario-validation';
 export * from './scenario-validation-error-messages';
